+++ frontend/src/entities/article/index.ts
@@ -5,5 +5,5 @@ export {
 	useCreateArticleMutation,
 	articlesApi,
 } from './api/articleApi';
-export { default as ArticlePreview } from './ui/ArticlePreview';
+export * from './ui';
 export { type TArticle } from './model/types';
