+++ frontend/src/entities/article/ui/ArticleMeta/ArticleMeta.module.css
@@ -1,0 +1,28 @@
+.articleMeta {
+	margin: 0 0 1rem 0;
+	display: flex;
+	align-items: center;
+	gap: 0.5rem;
+}
+
+.articleMeta img {
+	width: 32px;
+	aspect-ratio: 1;
+	border-radius: 50%;
+}
+
+.info {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+}
+
+.date {
+	color: hsl(var(--clr-gray));
+	font-size: var(--fs-300);
+	line-height: 1;
+}
+
+.actions_container {
+	margin-left: auto;
+}
