+++ frontend/src/entities/article/ui/ArticlePreview/ArticlePreview.module.css
@@ -3,35 +3,10 @@
 	padding: 1.5rem 0;
 }
 
-.articleMeta {
-	margin: 0 0 1rem 0;
-	display: flex;
-	align-items: center;
-	gap: 0.5rem;
-}
-
-.articleMeta img {
-	width: 32px;
-	aspect-ratio: 1;
-	border-radius: 50%;
-}
-
-.info {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-}
-
 .author {
 	color: hsl(var(--clr-primary));
 	text-decoration: none;
 	font-weight: 500;
-	line-height: 1;
-}
-
-.date {
-	color: hsl(var(--clr-gray));
-	font-size: var(--fs-300);
 	line-height: 1;
 }
 
@@ -73,10 +48,6 @@
 .tagList li {
 	display: inline-block;
 	margin-right: 3px;
-}
-
-.actions_container {
-	margin-left: auto;
 }
 
 .bottom {
