+++ frontend/src/entities/article/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as ArticleMeta } from './ArticleMeta/ArticleMeta';
+export { default as ArticlePreview } from './ArticlePreview/ArticlePreview';
+export { default as ArticlesList } from './ArticlesList/ArticlesList';
