+++ frontend/src/pages/articles/ui/ArticlesPage.module.css
@@ -207,8 +207,3 @@
 	display: inline-block;
 	margin-right: 3px;
 }
-
-.pagination_container {
-	margin-top: 1rem;
-	margin-bottom: 5rem;
-}
