+++ frontend/src/shared/ui/index.ts
@@ -2,3 +2,4 @@ export { default as Layout } from './Layout/Layout';
 export { default as Button } from './Button/Button';
 export { default as Input } from './Input/Input';
 export { default as Textarea } from './TextArea/Textarea';
+export { default as Pagination } from './Pagination/Pagination';
