+++ frontend/src/widgets/ArticlesList/index.ts
@@ -1,1 +1,1 @@
-export { default as ArticlesList } from './ui/ArticlesList';
+export { default as GlobalArticlesList } from './ui/GlobalArticlesList';
