+++ frontend/src/widgets/ArticlesList/ui/GlobalArticlesList.module.css
@@ -1,0 +1,4 @@
+.pagination_container {
+	margin-top: 1rem;
+	margin-bottom: 5rem;
+}
