--- frontend/src/widgets/Pagination/index.ts
@@ -1,1 +1,0 @@
-export { default as Pagination } from './ui/Pagination';
